+++ lib/Conversion/ConvVectorization/CMakeLists.txt
@@ -2,4 +2,5 @@ add_mlir_library(CBConvVectorization
   CBConvVectorization.cpp
   GEMMPointwiseConv2DNhwcHwcf.cpp
   CBPoolingVectorization.cpp
+  PoolingVectorization.cpp
   )
